# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.IntensificationRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  model.IntensificationResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  model.MuscleGroupRequest:
    properties:
      body_part:
        type: string
      description:
        type: string
      level:
        type: string
      name:
        type: string
    type: object
  model.MuscleGroupResponse:
    properties:
      body_part:
        type: string
      description:
        type: string
      id:
        type: integer
      level:
        type: string
      name:
        type: string
    type: object
...
  model.ProgramRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
      user_id:
        type: integer
    type: object
  model.ProgramResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for a sport application.
  title: Sportin API
  version: "1.0"
paths:
...
  /intensifications:
    get:
      consumes:
      - application/json
      description: Get all Intensification
      parameters:
      - description: Intensification object that needs to be created
        in: body
        name: Category
        required: true
        schema:
          $ref: '#/definitions/model.IntensificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.IntensificationResponse'
      summary: Get all Intensification
      tags:
      - Intensification
    post:
      consumes:
      - application/json
      description: Create a new Intensification
      parameters:
      - description: Exercise object that needs to be created
        in: body
        name: exercise
        required: true
        schema:
          $ref: '#/definitions/model.IntensificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.IntensificationResponse'
      summary: Create a new Intensification
      tags:
      - Intensification
  /intensifications/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Intensification
      parameters:
      - description: Intensification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete Intensification
      tags:
      - Intensification
    get:
      consumes:
      - application/json
      description: Get Intensification
      parameters:
      - description: Intensification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.IntensificationResponse'
      summary: Get Intensification
      tags:
      - Intensification
    put:
      consumes:
      - application/json
      description: Update Intensification
      parameters:
      - description: Intensification ID
        in: path
        name: id
        required: true
        type: integer
      - description: Intensification object that needs to be updated
        in: body
        name: intensification
        required: true
        schema:
          $ref: '#/definitions/model.IntensificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.IntensificationResponse'
      summary: Update Intensification
      tags:
      - Intensification
...
  /muscleGroup:
    get:
      consumes:
      - application/json
      description: Get all muscle groups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.MuscleGroupResponse'
            type: array
      summary: Get all muscle groups
      tags:
      - MuscleGroup
    post:
      consumes:
      - application/json
      description: Create a new muscle group
      parameters:
      - description: Muscle Group object that needs to be created
        in: body
        name: muscleGroup
        required: true
        schema:
          $ref: '#/definitions/model.MuscleGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MuscleGroupResponse'
      summary: Create a new muscle group
      tags:
      - MuscleGroup
  /muscleGroup/{id}:
    delete:
      consumes:
      - application/json
      description: Delete muscle group
      parameters:
      - description: Muscle Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete muscle group
      tags:
      - MuscleGroup
    get:
      consumes:
      - application/json
      description: Get muscle group
      parameters:
      - description: Muscle Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MuscleGroupResponse'
      summary: Get muscle group
      tags:
      - MuscleGroup
    put:
      consumes:
      - application/json
      description: Update muscle group
      parameters:
      - description: Muscle Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MuscleGroupResponse'
      summary: Update muscle group
      tags:
      - MuscleGroup
  /programs:
    get:
      consumes:
      - application/json
      description: Get all programs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ProgramResponse'
            type: array
      summary: Get all programs
      tags:
      - Program
    post:
      consumes:
      - application/json
      description: Create a new program
      parameters:
      - description: Program object that needs to be created
        in: body
        name: program
        required: true
        schema:
          $ref: '#/definitions/model.ProgramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProgramResponse'
      summary: Create a new program
      tags:
      - Program
  /programs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a program
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete a program
      tags:
      - Program
    get:
      consumes:
      - application/json
      description: Get program
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProgramResponse'
      summary: Get program
      tags:
      - Program
    put:
      consumes:
      - application/json
      description: Update a program
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: integer
      - description: Program object that needs to be updated
        in: body
        name: program
        required: true
        schema:
          $ref: '#/definitions/model.ProgramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProgramResponse'
      summary: Update a program
      tags:
      - Program
...
swagger: "2.0"